Move to C interface of HDF5
The C++ API is very unstable. As it is a very simple wrapper around the C API, we hope to solve at least part of the issue now.

This also fixes versions prior to hdf5-1.8.15, where the C++ API contained a reference counter bug. We claimed that the simulator used to work with versions 1.8.12--1.8.15, but while it did compile, it did not run because of this bug in HDF5.

Because we want to use H5free_memory(), the minimum required version is now bumped to 1.8.13.